+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/StatesMachine.kt
@@ -13,13 +13,53 @@ private suspend fun <I : State> StatesMachine.launchStateHandling(
     }
 }
 
-class StatesMachine (
+/**
+ * Default [StatesMachine] may [startChain] and use inside logic for handling [State]s. By default you may use
+ * [DefaultStatesMachine] or build it with [dev.inmo.micro_utils.fsm.common.dsl.buildFSM]. Implementers MUST NOT start
+ * handling until [start] method will be called
+ */
+interface StatesMachine : StatesHandler<State> {
+    /**
+     * Starts handling of [State]s
+     */
+    fun start(scope: CoroutineScope): Job
+
+    /**
+     * Start chain of [State]s witn [state]
+     */
+    suspend fun startChain(state: State)
+
+    companion object {
+        /**
+         * Creates [DefaultStatesMachine]
+         */
+        operator fun invoke(
+            statesManager: StatesManager,
+            handlers: List<StateHandlerHolder<*>>
+        ) = DefaultStatesMachine(statesManager, handlers)
+    }
+}
+
+/**
+ * Default realization of [StatesMachine]. It uses [statesManager] for incapsulation of [State]s storing and contexts
+ * resolving, and uses [launchStateHandling] for [State] handling
+ */
+class DefaultStatesMachine (
     private val statesManager: StatesManager,
     private val handlers: List<StateHandlerHolder<*>>
-) : StatesHandler<State> {
+) : StatesMachine {
+    /**
+     * Will call [launchStateHandling] for state handling
+     */
     override suspend fun StatesMachine.handleState(state: State): State? = launchStateHandling(state, handlers)
 
-    fun start(scope: CoroutineScope): Job = scope.launchSafelyWithoutExceptions {
+    /**
+     * Launch handling of states. On [statesManager] [StatesManager.onStartChain],
+     * [statesManager] [StatesManager.onChainStateUpdated] will be called lambda with performing of state. If
+     * [launchStateHandling] will returns some [State] then [statesManager] [StatesManager.update] will be used, otherwise
+     * [StatesManager.endChain].
+     */
+    override fun start(scope: CoroutineScope): Job = scope.launchSafelyWithoutExceptions {
         val statePerformer: suspend (State) -> Unit = { state: State ->
             val newState = launchStateHandling(state, handlers)
             if (newState != null) {
@@ -40,7 +80,10 @@ class StatesMachine (
         }
     }
 
-    suspend fun startChain(state: State) {
+    /**
+     * Just calls [StatesManager.startChain] of [statesManager]
+     */
+    override suspend fun startChain(state: State) {
         statesManager.startChain(state)
     }
 }